Add test that ensures all commands have help text
I realised while adding the /invite command that it’s easy
to forget to add help text for a new command. This presumes
that all working commands will have a corresponding file
within the commands directory, but that is currently true.

If/when localisation happens, this should be extended to
check every locale file.