Add Coveralls (#22)
Maybe there’s a better way, but this would have [revealed](https://coveralls.io/builds/12790474) without need for direct experience that deletion has no coverage (#17) and that `UnsubscribedTargetResponse` doesn’t work (#19).

I vaguely recall that responses are one of the lacking parts of the tests, so maybe the build should fail if coverage is missing for every response class? Not sure about that, so I’ll think more about whether that makes sense.